+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+### Removed
+- removed the Auth class as it is no longer used
+
 ## [2.3.7] - Logo & PDF lib
 ### Updated
 - updated logo used for manifest