+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [Unreleased]
+### Removed
+- removed the Auth class as it is no longer used
+
+### Changed
+- moved location API url from class Shipment to class PickupPoints
+- changed locations API url to v3
+
+### Added
+- added request error return
+
 ## [2.3.7] - Logo & PDF lib
 ### Updated
 - updated logo used for manifest